+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [Unreleased]
+
+## v0.1.11
+
+- Added a conflict check for ag2 and autogen-agentchat
+- Updated exporting skills: create a new file the the skill secrets
+- Updated ag2 to 0.4.1
+- Other dependency updates
+
 ## v0.1.10
 
 - Fixed an issue with using a chromadb client without telemetry